# Code generator performance improvements