This makes everything faster, but it in particular makes things faster if you're using `compact: true`

Here's some unscientific numbers from me parsing an AST then generating code from it 100 times:

Master:

| Compact	| SourceMaps| Comments	| Time  |
| --------- | --------- | --------- | ----- |
| true		| false		| false		| 4.15s |
| true		| false		| true		| 4.11s |
| true		| true		| false		| 6.67s |
| true		| true		| true		| 6.92s |
| false		| false		| false		| 5.65s |
| false		| false		| true		| 5.64s |
| false		| true		| false		| 8.72s |
| false		| true		| true		| 8.71s |


With PR:

| Compact	| SourceMaps| Comments	| Time  |
| --------- | --------- | --------- | ----- |
| true		| false		| false		| 1.35s |
| true		| false		| true		| 1.26s |
| true		| true		| false		| 3.02s |
| true		| true		| true		| 3.23s |
| false		| false		| false		| 3.17s |
| false		| false		| true		| 3.23s |
| false		| true		| false		| 5.42s |
| false		| true		| true		| 5.32s |