@loganfsmyth 2016-04-20
@samccone Basically doing
```
var content = fs.readFileSync(' <a 147KB file from our codebase >', 'utf8');
var ast = babylon.parse(content, {sourceType: 'module'});

var minified = false;
var sourceMaps = false;
var comments = false;
function gen(){
    return generator.default(ast, {sourceMaps: sourceMaps, sourceFileName: 'omg.js', minified: minified, comments: comments}, content).code;
}

// Run it a few times to get V8 started
console.log(gen().length);
console.log(gen().length);
console.log(gen().length);

var start = process.hrtime();

for (var i = 0; i < 100; i++){
    total += gen().length;
}

var end = process.hrtime(start);
console.log('time:', end[0] + end[1] / 1e9);
console.log(total);
```